added namespace and renamed to Semver
