+++ src/index.ts
@@ -11,4 +11,4 @@ export * from './Metadata'
 export * from './responses'
 export * from './Billing'
 export * from './LRUCache'
-export * from './AppIds'
+export * from './Semver'
